+++ template/RELEASE_NOTES.md
@@ -62,6 +62,7 @@
 * ANDDEP-1137 There is no support for FreeMarker templates in AndroidStudio 4.1, so file templates were modified in order to use them with Geminio library. Geminio compatible file templates are added to project template.
 * ANDDEP-1126 Fixed MigrationModule.kt, get version code from packageInfo.
 * ANDDEP-1138 Added supporting View Binding
+* ANDDEP-1144 Fix navigation
 ## 0.4.0
 ##### Template
 * Updated navigation mechanisms in template